centos-ci: add pynfs test job
Reserves two machines, one for the server and one for the client. After
configuring the server, the tests are compiled and run on the client.

A new job will automatically start when the nightly NFS-Ganesha RPMs are
available. The test only uses CentOS-7/x86_64 by default.

Results in Jenkins are currently incorrect. The test-script in client.sh
always returns success when some tests were run. This needs to get
addressed in the uptream pynfs repository.

Suggested-by: Jiffin Tony Thottan <jthottan@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>