Update CMake invocation for new goptions
CMake usage for options changed in Ganesha, so that it tracks default-on
verses explicitly enabled options, failing in the latter case when
dependencies are not present.

Change the CI builds as follows:

The main build CI now builds with an explicit request for GLUSTER, CEPH,
RGW, DBUS and so on.

The actual test runs only build with GLUSTER required, and the rest are
defaults.  This should allow build verification with minimal
interruption of testing.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>